+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog of threedi-modelchecker
 2.6.2 (unreleased)
 ------------------
 
+- Add warning check (0616) for surfaces for which no inflow is generated because of the surface table not being referred to in global settings.
+- Add warning check (0617) to warn if the surface table referred to in global settings is empty and no inflow will be generated for it.
 - Remove warning check 0029.